Add support for pre-install scripts
1. Extraction will be made to a temporary location side-by-side of the target directory
2. Pre-install scripts are run from the temporary location
3. Move the temporary directory to the target location
4. Post-install scripts are executed from the target location

The purpose of this is to have a chance to stop running instances of a
previous version of the application before extraction occurs.
This solves the problem when extraction cannot occur due to file locks in IIS and such.

A new property of the scripts object in the kolli.json file defines the list of preinstall scripts.
Other than the temporary location the preinstall scripts are executed just like existing postinstall scripts.